This is  the first commit in my HPQ project, on the master branch.
I have started by making use of my Liber Engine dll, which creates a 3D environment for visible shapes to exist in, complete with options for FOV, camera position and zoom, aswell as an array of mathematical structures used to create these shapes, such as vectors and points. This library was handmade by me in my last project.

I've already gotten started using a hierarchy of classes 'formula' and 'term' and so on to make sense of user input, process data and output.

There is also an experiment for creating panels that draw custom text when focused.